power-dashboard/requirements.txt
streamlit>=1.35.0
supabase>=2.4.0
pandas>=2.2.0
plotly>=5.22.0
python-dotenv>=1.0.0
